+++ src/Markdown/Extension/Figure/Node/Figure.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Jidaikobo\Markdown\Extension\Node;
+namespace Jidaikobo\Markdown\Extension\Figure\Node;
 
 use League\CommonMark\Node\Block\AbstractBlock;