+++ src/Markdown/Extension/Table/Node/TableCaption.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Jidaikobo\Markdown\Extension\Node;
+namespace Jidaikobo\Markdown\Extension\Table\Node;
 
 use League\CommonMark\Node\Block\AbstractBlock;