+++ src/Markdown/Extension/Table/Renderer/TableCaptionRenderer.php
@@ -2,20 +2,20 @@
 
 declare(strict_types=1);
 
-namespace Jidaikobo\Markdown\Extension\Renderer;
+namespace Jidaikobo\Markdown\Extension\Table\Renderer;
 
-use Jidaikobo\Markdown\Extension\Node\Figcaption;
+use Jidaikobo\Markdown\Extension\Table\Node\TableCaption;
 use League\CommonMark\Node\Node;
 use League\CommonMark\Renderer\ChildNodeRendererInterface;
 use League\CommonMark\Renderer\NodeRendererInterface;
 use League\CommonMark\Util\HtmlElement;
 
-final class FigcaptionRenderer implements NodeRendererInterface
+final class TableCaptionRenderer implements NodeRendererInterface
 {
     public function render(Node $node, ChildNodeRendererInterface $childRenderer): \Stringable
     {
-        Figcaption::assertInstanceOf($node);
+        TableCaption::assertInstanceOf($node);
 
-        return new HtmlElement('figcaption', [], $childRenderer->renderNodes($node->children()));
+        return new HtmlElement('caption', [], $childRenderer->renderNodes($node->children()));
     }
 }